track local validators
define LocalValidatorRegistry to track indices of validators that have
registered with the node.

read indices of local validators from LocalValidatorRegistry when updating
validator custody requirements.

borrow logic from ValidatorMonitor for registering and pruning validators from
the map in LocalValidatorRegistry.

register validators and prune validators in tandem from LocalValidatorRegistry and ValidatorMonitor.
TODO: re-use validator registry in ValidatorMonitor.

we need to be able to determine the indices of connected validators even if
metrics are disabled.