chore: fix lints, types and unit tests in peerDAS (#7707)
**Motivation**

It'd be nice to get tests and lints passing on the peerDAS branch, both
to get it ready to merge into unstable and so we can check PRs into the
peerDAS branch.

**Description**

The only intended functional change is to `initPeerIdAndEnr`, will add a
comment on that change. It might be easier to review commit by commit,
to more easily view the lint fixes separately from the unit test fixes.